+++ service/src/test/java/uk/nhs/adaptors/gp2gp/ehr/EhrExtractStatusServiceTest.java
@@ -1,10 +1,8 @@
 package uk.nhs.adaptors.gp2gp.ehr;
 
-import org.bson.Document;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -13,10 +11,8 @@ import org.springframework.data.mongodb.core.FindAndModifyOptions;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.core.query.Update;
-import org.springframework.data.mongodb.core.query.UpdateDefinition;
 import org.springframework.test.util.ReflectionTestUtils;
 import uk.nhs.adaptors.gp2gp.common.service.TimestampService;
-import uk.nhs.adaptors.gp2gp.ehr.exception.EhrExtractException;
 import uk.nhs.adaptors.gp2gp.ehr.model.EhrExtractStatus;
 import uk.nhs.adaptors.gp2gp.mhs.exception.UnrecognisedInteractionIdException;
 import java.time.Duration;
@@ -25,7 +21,6 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -33,9 +28,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
@@ -49,15 +42,7 @@ class EhrExtractStatusServiceTest {
     public static final int EIGHT_DAYS = 8;
     public static final String ERROR_CODE = "99";
     public static final String ALTERNATIVE_ERROR_CODE = "26";
-    public static final String ERROR_MESSAGE = "The acknowledgement has been received after 8 days";
-    private static final Instant NOW = Instant.now();
-    private static final Instant FIVE_DAYS_AGO = NOW.minus(Duration.ofDays(5));
-    public static final String ACK_TYPE = "AA";
-    public static final int TWENTY_DAYS = 20;
-
-    private ArgumentCaptor<Query> queryCaptor = ArgumentCaptor.forClass(Query.class);
-    private ArgumentCaptor<Update> updateCaptor = ArgumentCaptor.forClass(Update.class);
-    private ArgumentCaptor<Class<EhrExtractStatus>> classCaptor = ArgumentCaptor.forClass(Class.class);
+    public static final String ERROR_MESSAGE = "No acknowledgement has been received within 8 days";
 
     @Mock
     private Logger logger;
@@ -113,7 +98,9 @@ class EhrExtractStatusServiceTest {
         ehrExtractStatusServiceSpy.updateEhrExtractStatusAck(conversationId, ack);
 
         verify(logger, times(1))
-            .warn("Received an ACK message with a conversation_id: {}, but it will be ignored", conversationId);
+            .warn("Received an ACK message with conversation_id: {}, "
+                  + "but it is being ignored because the EhrExtract has already been marked as a failure and 8 days have passed.",
+                  conversationId);
     }
 
     @Test
@@ -291,7 +278,7 @@ class EhrExtractStatusServiceTest {
 
         doReturn(ehrExtractStatusWithEhrReceivedAckWithErrors).when(ehrExtractStatusRepository).findByConversationId(conversationId);
 
-        boolean result = ehrExtractStatusServiceSpy.hasEhrStatusReceivedAckWithUnexpectedConditionErrors(conversationId);
+        boolean result = ehrExtractStatusServiceSpy.hasErrorsInEhrStatusReceivedAck(conversationId);
 
         assertFalse(result);
     }
@@ -316,7 +303,7 @@ class EhrExtractStatusServiceTest {
 
         doReturn(ehrExtractStatusWithEhrReceivedAckWithErrors).when(ehrExtractStatusRepository).findByConversationId(conversationId);
 
-        assertFalse(ehrExtractStatusServiceSpy.hasEhrStatusReceivedAckWithUnexpectedConditionErrors(conversationId));
+        assertFalse(ehrExtractStatusServiceSpy.hasErrorsInEhrStatusReceivedAck(conversationId));
     }
 
     @Test
@@ -339,7 +326,7 @@ class EhrExtractStatusServiceTest {
 
         doReturn(ehrExtractStatusWithEhrReceivedAckWithErrors).when(ehrExtractStatusRepository).findByConversationId(conversationId);
 
-        assertFalse(ehrExtractStatusServiceSpy.hasEhrStatusReceivedAckWithUnexpectedConditionErrors(conversationId));
+        assertFalse(ehrExtractStatusServiceSpy.hasErrorsInEhrStatusReceivedAck(conversationId));
     }
 
     @Test
@@ -362,7 +349,7 @@ class EhrExtractStatusServiceTest {
 
         doReturn(ehrExtractStatusWithEhrReceivedAckWithErrors).when(ehrExtractStatusRepository).findByConversationId(conversationId);
 
-        assertFalse(ehrExtractStatusServiceSpy.hasEhrStatusReceivedAckWithUnexpectedConditionErrors(conversationId));
+        assertFalse(ehrExtractStatusServiceSpy.hasErrorsInEhrStatusReceivedAck(conversationId));
     }
 
     @Test
@@ -390,7 +377,7 @@ class EhrExtractStatusServiceTest {
 
         doReturn(ehrExtractStatusWithEhrReceivedAckWithErrors).when(ehrExtractStatusRepository).findByConversationId(conversationId);
 
-        assertDoesNotThrow(() -> ehrExtractStatusServiceSpy.hasEhrStatusReceivedAckWithUnexpectedConditionErrors(conversationId));
+        assertDoesNotThrow(() -> ehrExtractStatusServiceSpy.hasErrorsInEhrStatusReceivedAck(conversationId));
     }
 
     @Test
@@ -418,200 +405,9 @@ class EhrExtractStatusServiceTest {
 
         doReturn(ehrExtractStatusWithEhrReceivedAckWithErrors).when(ehrExtractStatusRepository).findByConversationId(conversationId);
 
-        boolean result = ehrExtractStatusServiceSpy.hasEhrStatusReceivedAckWithUnexpectedConditionErrors(conversationId);
+        boolean result = ehrExtractStatusServiceSpy.hasErrorsInEhrStatusReceivedAck(conversationId);
 
         assertTrue(result);
-    }
-
-    @Test
-    void shouldUpdateStatusWithErrorWhenEhrExtractAckTimeoutOccurs() {
-
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-        EhrExtractStatus ehrExtractStatusUpdated = EhrExtractStatus.builder().build();
-
-        doReturn(List.of(ehrExtractStatus)).when(ehrExtractStatusServiceSpy).findInProgressTransfers();
-        doReturn(ehrExtractStatusUpdated).when(mongoTemplate).findAndModify(any(Query.class), any(UpdateDefinition.class),
-                                                         any(FindAndModifyOptions.class), any());
-        when(ehrExtractStatusServiceSpy.logger()).thenReturn(logger);
-
-        ehrExtractStatusServiceSpy.checkForEhrExtractAckTimeouts();
-
-        verify(ehrExtractStatusServiceSpy, times(1))
-            .updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(any(Stream.class),
-                                                                           eq(ERROR_CODE),
-                                                                           eq(ERROR_MESSAGE));
-        verify(logger).info("EHR status (EHR received acknowledgement) record successfully "
-                               + "updated in the database with error information conversation_id: {}", inProgressConversationId);
-    }
-
-    @Test
-    void shouldUpdateStatusWithErrorAndSpecificErrorCodeAndMessage() {
-
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-        EhrExtractStatus ehrExtractStatusUpdated = EhrExtractStatus.builder().build();
-
-        doReturn(ehrExtractStatusUpdated).when(mongoTemplate).findAndModify(any(Query.class), any(UpdateDefinition.class),
-                                                                            any(FindAndModifyOptions.class), any());
-        when(ehrExtractStatusServiceSpy.logger()).thenReturn(logger);
-
-        ehrExtractStatusServiceSpy.updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(Stream.of(ehrExtractStatus),
-                                                                                                 ERROR_CODE,
-                                                                                                 ERROR_MESSAGE);
-
-        verify(logger).info("EHR status (EHR received acknowledgement) record successfully "
-                            + "updated in the database with error information conversation_id: {}", inProgressConversationId);
-        verify(mongoTemplate, times(1)).findAndModify(queryCaptor.capture(),
-                                                                            updateCaptor.capture(),
-                                                                            any(FindAndModifyOptions.class),
-                                                                            classCaptor.capture());
-
-        assertEquals(ERROR_CODE, ((EhrExtractStatus.EhrReceivedAcknowledgement.ErrorDetails) ((Document) updateCaptor.getValue()
-                                                                                .getUpdateObject()
-                                                                                .get("$addToSet"))
-                                                                                .get("ehrReceivedAcknowledgement.errors")).getCode());
-        assertEquals(ERROR_MESSAGE, ((EhrExtractStatus.EhrReceivedAcknowledgement.ErrorDetails) ((Document) updateCaptor.getValue()
-                                                                                .getUpdateObject()
-                                                                                .get("$addToSet"))
-                                                                                .get("ehrReceivedAcknowledgement.errors")).getDisplay());
-    }
-
-    @Test
-    void shouldNotUpdateStatusWhenEhrReceivedAcknowledgementIsNotNull() {
-
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-        ehrExtractStatus.setEhrReceivedAcknowledgement(EhrExtractStatus.EhrReceivedAcknowledgement.builder().build());
-
-        doReturn(List.of(ehrExtractStatus)).when(ehrExtractStatusServiceSpy).findInProgressTransfers();
-
-        ehrExtractStatusServiceSpy.checkForEhrExtractAckTimeouts();
-
-        verify(ehrExtractStatusServiceSpy, never())
-            .updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(Stream.of(ehrExtractStatus), ERROR_CODE, ERROR_MESSAGE);
-    }
-
-    @Test
-    void shouldNotUpdateStatusWhenNoInProgressTransfersExist() {
-
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-
-        doReturn(List.of()).when(ehrExtractStatusServiceSpy).findInProgressTransfers();
-
-        ehrExtractStatusServiceSpy.checkForEhrExtractAckTimeouts();
-
-        verify(ehrExtractStatusServiceSpy, never())
-            .updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(Stream.of(), ERROR_CODE, ERROR_MESSAGE);
-    }
-
-    @Test
-    void shouldNotUpdateStatusWhenInProgressTransfersWithNullEhrExtractCorePendingExist() {
-
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-        ehrExtractStatus.setEhrExtractCorePending(null);
-
-        doReturn(List.of(ehrExtractStatus)).when(ehrExtractStatusServiceSpy).findInProgressTransfers();
-
-        ehrExtractStatusServiceSpy.checkForEhrExtractAckTimeouts();
-
-        verify(ehrExtractStatusServiceSpy, never())
-            .updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(Stream.of(), ERROR_CODE, ERROR_MESSAGE);
-    }
-
-    @Test
-    void shouldNotDoAnythingIfThereAreNoEhrExtractStatusThatExceeded8Days() {
-
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        ReflectionTestUtils.setField(ehrExtractStatusServiceSpy, "ehrExtractSentDaysLimit", EIGHT_DAYS);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-        ehrExtractStatus.setEhrExtractCorePending(
-            EhrExtractStatus.EhrExtractCorePending
-                .builder()
-                .sentAt(FIVE_DAYS_AGO)
-                .build());
-
-        doReturn(List.of(ehrExtractStatus)).when(ehrExtractStatusServiceSpy).findInProgressTransfers();
-
-        ehrExtractStatusServiceSpy.checkForEhrExtractAckTimeouts();
-
-        verify(ehrExtractStatusServiceSpy, never())
-            .updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(Stream.of(), ERROR_CODE, ERROR_MESSAGE);
-    }
-
-    @Test
-    void updateEhrExtractStatusListWithEhrReceivedAcknowledgementError() {
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-
-        doReturn(null).when(mongoTemplate).findAndModify(any(Query.class), any(UpdateDefinition.class),
-                                                                    any(FindAndModifyOptions.class), any());
-        when(ehrExtractStatusServiceSpy.logger()).thenReturn(logger);
-
-        var exception = assertThrows(EhrExtractException.class,
-                                     () -> ehrExtractStatusServiceSpy.updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(
-                                         Stream.of(ehrExtractStatus), ERROR_CODE, ERROR_MESSAGE));
-
-        assertEquals("Couldn't update EHR received acknowledgement with error information because EHR status doesn't exist, "
-                     + "conversation_id: " + inProgressConversationId,
-                     exception.getMessage());
-        verify(logger).error(eq("An error occurred when updating EHR Extract with Ack erorrs, EHR Extract Status conversation_id: {}"),
-                             eq(inProgressConversationId),
-                             any(EhrExtractException.class));
-    }
-
-    @Test
-    void shouldCatchExceptionIfUnexpectedConditionAriseWhileUpdatingEhrExtractStatusListWithEhrReceivedAcknowledgementError() {
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-
-        doThrow(new NullPointerException())
-            .when(mongoTemplate).findAndModify(any(Query.class), any(UpdateDefinition.class), any(FindAndModifyOptions.class), any());
-        when(ehrExtractStatusServiceSpy.logger()).thenReturn(logger);
-
-        assertThrows(Exception.class,
-                         () -> ehrExtractStatusServiceSpy.updateEhrExtractStatusListWithEhrReceivedAcknowledgementError(
-                             Stream.of(ehrExtractStatus), ERROR_CODE, ERROR_MESSAGE));
-
-        verify(logger).error(eq("An unexpected error occurred for conversation_id: {}"),
-                             eq(inProgressConversationId),
-                             any(NullPointerException.class));
-    }
-
-    @Test
-    void whenEhrExtractStatusIsNullInterceptExceptionAndLogErrorMsg() {
-
-        EhrExtractStatusService ehrExtractStatusServiceSpy = spy(ehrExtractStatusService);
-        var inProgressConversationId = generateRandomUppercaseUUID();
-
-        EhrExtractStatus ehrExtractStatus = addInProgressTransfers(inProgressConversationId);
-
-        doReturn(List.of(ehrExtractStatus)).when(ehrExtractStatusServiceSpy).findInProgressTransfers();
-        doReturn(null).when(mongoTemplate).findAndModify(any(Query.class), any(UpdateDefinition.class),
-                                                         any(FindAndModifyOptions.class), any());
-        when(ehrExtractStatusServiceSpy.logger()).thenReturn(logger);
-
-        var exception = assertThrows(EhrExtractException.class, () -> ehrExtractStatusServiceSpy.checkForEhrExtractAckTimeouts());
-
-        assertEquals("Couldn't update EHR received acknowledgement with error information because EHR status doesn't exist, "
-                     + "conversation_id: " + inProgressConversationId,
-                     exception.getMessage());
-        verify(logger).error(eq("An error occurred when updating EHR Extract with Ack erorrs, EHR Extract Status conversation_id: {}"),
-                             eq(inProgressConversationId),
-                             any(EhrExtractException.class));
     }
 
     @Test
@@ -652,60 +448,6 @@ class EhrExtractStatusServiceTest {
         return UUID.randomUUID().toString().toUpperCase();
     }
 
-    private EhrExtractStatus addInProgressTransfers(String conversationId) {
-        EhrExtractStatus extractStatus = EhrExtractStatus.builder()
-            .ackPending(buildPositiveAckPending())
-            .ackToRequester(buildPositiveAckToRequester())
-            .conversationId(conversationId)
-            .created(Instant.now().minus(Duration.ofDays(TWENTY_DAYS)))
-            .ehrExtractCore(EhrExtractStatus.EhrExtractCore.builder()
-                                .sentAt(Instant.now().minus(Duration.ofDays(NINE_DAYS)))
-                                .build())
-            .ehrExtractCorePending(EhrExtractStatus.EhrExtractCorePending.builder()
-                                       .sentAt(Instant.now().minus(Duration.ofDays(NINE_DAYS)))
-                                       .taskId(generateRandomUppercaseUUID())
-                                       .build())
-            .ehrExtractMessageId(generateRandomUppercaseUUID())
-            .ehrRequest(buildEhrRequest())
-            .gpcAccessDocument(EhrExtractStatus.GpcAccessDocument.builder()
-                                   .documents(List.of())
-                                   .build())
-            .gpcAccessStructured(EhrExtractStatus.GpcAccessStructured.builder()
-                                     .accessedAt(Instant.now().minus(Duration.ofDays(NINE_DAYS)))
-                                     .objectName(generateRandomUppercaseUUID() + ".json")
-                                     .taskId(generateRandomUppercaseUUID())
-                                     .build())
-            .messageTimestamp(Instant.now().minus(Duration.ofDays(NINE_DAYS)))
-            .updatedAt(Instant.now().minus(Duration.ofDays(NINE_DAYS)))
-            .build();
 
-        ehrExtractStatusRepository.save(extractStatus);
-        return extractStatus;
-    }
-
-    private EhrExtractStatus.AckPending buildPositiveAckPending() {
-        return EhrExtractStatus.AckPending.builder()
-            .messageId(generateRandomUppercaseUUID())
-            .taskId(generateRandomUppercaseUUID())
-            .typeCode(ACK_TYPE)
-            .updatedAt(FIVE_DAYS_AGO.toString())
-            .build();
-    }
-
-    private EhrExtractStatus.AckToRequester buildPositiveAckToRequester() {
-        return EhrExtractStatus.AckToRequester.builder()
-            .detail(null)
-            .messageId(generateRandomUppercaseUUID())
-            .reasonCode(null)
-            .taskId(generateRandomUppercaseUUID())
-            .typeCode(ACK_TYPE)
-            .build();
-    }
-
-    private EhrExtractStatus.EhrRequest buildEhrRequest() {
-        return EhrExtractStatus.EhrRequest.builder()
-            .messageId(generateRandomUppercaseUUID())
-            .build();
-    }
 
 }
